Fix seating page font size on zoom and small screens
