+++ src/components/pages/SeatLayout.js
@@ -13,11 +13,12 @@ import FirstRow from "./movie_detail_page/FirstRow";
 import SeatingChart from "../seating_page/SeatingChart";
 import ShowTime from "../seating_page/ShowTime";
 import ShowDate from "../seating_page/ShowDate";
-import Theater from "../seating_page/Theater";
+import TheaterRoomName from "../seating_page/TheaterRoomName";
 import RuntimeElement from "../seating_page/RuntimeElement";
 import ScreenLine from "../seating_page/ScreenLine";
 import Legends from "../seating_page/Legends";
 import ReserveSeatButton from "../seating_page/ReserveSeatButton";
+import "../seating_page/TheaterTextSizing.css";
 
 const SeatLayout = (props) => {
     let screeningId = getIdFromUrl();
@@ -97,9 +98,9 @@ const SeatLayout = (props) => {
                                         backdrop={backdrop}/>
                     </div>
                     <div>
-                        <h1 style={titleStyle}>{title} <span className="hazy">({releaseDate.substring(0, 4)})</span>
+                        <h1 className="movie-title">{title} <span className="hazy">({releaseDate.substring(0, 4)})</span>
                         </h1>
-                        {room ? <Theater theater={room.name}/> : <div/>}
+                        {room ? <TheaterRoomName theater={room.name}/> : <div/>}
                         <ShowTime time={startingTime}/>
                         <ShowDate date={startingDate}/>
                         <RuntimeElement runtime={runtime}/>
@@ -125,15 +126,5 @@ const mainColumnStyle = {
     flexFlow: "row wrap",
     height: "1500px",
     padding: "0"
-}
-
-const titleStyle = {
-    position: "absolute",
-    fontSize: "3em",
-    top: "30%",
-    left: "53%",
-    transform: "translate(-50%, -7%)",
-    zIndex: "1",
-    color: "white",
 }
