+++ src/components/seating_page/ShowDate.js
@@ -1,11 +1,14 @@
 import React from "react";
+import "./TheaterTextSizing.css";
 
 const ShowDate = (props) => {
 
     return (
         <div style={dateStyle}>
-            <div className="hazy" style={{fontSize: "1.2em"}}>Date</div>
-            <div style={{fontSize: "2.5em"}}>{props.date}</div>
+            {/*<div className="hazy" style={{fontSize: "1.2em"}}>Date</div>*/}
+            <div className="hazy sub-title">Date</div>
+            {/*<div style={{fontSize: "2.5em"}}>{props.date}</div>*/}
+            <div className="field-data">{props.date}</div>
         </div>
     )
 }
